+++ packages/openauth/src/client.ts
@@ -15,19 +15,37 @@ import {
 } from "./error.js"
 import { generatePKCE } from "./pkce.js"
 
+/**
+ * The well-known information for an OAuth 2.0 authorization server.
+ */
 export interface WellKnown {
+  /**
+   * The URI to the JWKS endpoint.
+   */
   jwks_uri: string
+  /**
+   * The URI to the token endpoint.
+   */
   token_endpoint: string
+  /**
+   * The URI to the authorization endpoint.
+   */
   authorization_endpoint: string
 }
 
+/**
+ * The tokens returned by the authorization server.
+ */
 export interface Tokens {
+  /**
+   * The access token.
+   */
   access: string
+  /**
+   * The refresh token.
+   */
   refresh: string
 }
-
-const jwksCache = new Map<string, ReturnType<typeof createLocalJWKSet>>()
-const issuerCache = new Map<string, WellKnown>()
 
 interface ResponseLike {
   json(): Promise<unknown>
@@ -35,11 +53,134 @@ interface ResponseLike {
 }
 type FetchLike = (...args: any[]) => Promise<ResponseLike>
 
-export function createClient(input: {
+export type Challenge = {
+  state: string
+  verifier?: string
+}
+
+/**
+ * The input object for the client.
+ */
+export interface ClientInput {
+  /**
+   * The client ID.
+   */
   clientID: string
+  /**
+   * The issuer URL.
+   */
   issuer?: string
+  /**
+   * The fetch function to use.
+   */
   fetch?: FetchLike
-}) {
+}
+
+export interface AuthorizeOptions {
+  pkce?: boolean
+  provider?: string
+}
+
+export interface AuthorizeResult {
+  challenge: Challenge
+  url: string
+}
+
+export interface ExchangeSuccess {
+  err: false
+  tokens: Tokens
+}
+
+export interface ExchangeError {
+  err: InvalidAuthorizationCodeError
+}
+
+export interface RefreshOptions {
+  access?: string
+}
+
+export interface RefreshSuccess {
+  err: false
+  tokens?: Tokens
+}
+
+export interface RefreshError {
+  err: InvalidRefreshTokenError | InvalidAccessTokenError
+}
+
+export interface VerifyOptions {
+  refresh?: string
+  issuer?: string
+  audience?: string
+  fetch?: typeof fetch
+}
+
+export interface VerifyResult<T extends SubjectSchema> {
+  err?: undefined
+  tokens?: Tokens
+  aud: string
+  subject: {
+    [type in keyof T]: { type: type; properties: v1.InferOutput<T[type]> }
+  }[keyof T]
+}
+
+export interface VerifyError {
+  err: InvalidRefreshTokenError | InvalidAccessTokenError
+}
+
+export interface Client {
+  /**
+   * Authorize the client.
+   * @param redirectURI - The redirect URI.
+   * @param response - The response type.
+   * @param opts - Authorization options.
+   */
+  authorize(
+    redirectURI: string,
+    response: "code" | "token",
+    opts?: AuthorizeOptions,
+  ): Promise<AuthorizeResult>
+  /**
+   * Exchange the authorization code for tokens.
+   * @param code - The authorization code.
+   * @param redirectURI - The redirect URI.
+   * @param verifier - The verifier.
+   */
+  exchange(
+    code: string,
+    redirectURI: string,
+    verifier?: string,
+  ): Promise<ExchangeSuccess | ExchangeError>
+  /**
+   * Refresh the tokens.
+   * @param refresh - The refresh token.
+   * @param opts - Refresh options.
+   */
+  refresh(
+    refresh: string,
+    opts?: RefreshOptions,
+  ): Promise<RefreshSuccess | RefreshError>
+  /**
+   * Verify the token.
+   * @param subjects - The subjects.
+   * @param token - The token.
+   * @param options - Verification options.
+   */
+  verify<T extends SubjectSchema>(
+    subjects: T,
+    token: string,
+    options?: VerifyOptions,
+  ): Promise<VerifyResult<T> | VerifyError>
+}
+
+/**
+ * Create a client object for interacting with an OAuth 2.0 authorization server.
+ * @param input - The input object containing the client ID, issuer, and optional fetch function.
+ * @returns An object containing methods for authorizing, exchanging tokens, refreshing tokens, and verifying tokens.
+ */
+export function createClient(input: ClientInput): Client {
+  const jwksCache = new Map<string, ReturnType<typeof createLocalJWKSet>>()
+  const issuerCache = new Map<string, WellKnown>()
   const issuer = input.issuer || process.env.OPENAUTH_ISSUER
   if (!issuer) throw new Error("No issuer")
   const f = input.fetch ?? fetch
@@ -67,20 +208,34 @@ export function createClient(input: {
   }
 
   const result = {
-    authorize(
+    async authorize(
       redirectURI: string,
       response: "code" | "token",
-      opts?: {
-        provider?: string
-      },
+      opts?: AuthorizeOptions,
     ) {
       const result = new URL(issuer + "/authorize")
-      if (opts?.provider) result.searchParams.set("provider", opts.provider)
+      const challenge: Challenge = {
+        state: crypto.randomUUID(),
+      }
       result.searchParams.set("client_id", input.clientID)
       result.searchParams.set("redirect_uri", redirectURI)
       result.searchParams.set("response_type", response)
-      return result.toString()
+      result.searchParams.set("state", challenge.state)
+      if (opts?.provider) result.searchParams.set("provider", opts.provider)
+      if (opts?.pkce && response === "code") {
+        const pkce = await generatePKCE()
+        result.searchParams.set("code_challenge_method", "S256")
+        result.searchParams.set("code_challenge", pkce.challenge)
+        challenge.verifier = pkce.verifier
+      }
+      return {
+        challenge,
+        url: result.toString(),
+      }
     },
+    /**
+     * @deprecated use `authorize` instead, it will do pkce by default unless disabled with `opts.pkce = false`
+     */
     async pkce(
       redirectURI: string,
       opts?: {
@@ -97,7 +252,11 @@ export function createClient(input: {
       result.searchParams.set("code_challenge", pkce.challenge)
       return [pkce.verifier, result.toString()]
     },
-    async exchange(code: string, redirectURI: string, verifier?: string) {
+    async exchange(
+      code: string,
+      redirectURI: string,
+      verifier?: string,
+    ): Promise<ExchangeSuccess | ExchangeError> {
       const tokens = await f(issuer + "/token", {
         method: "POST",
         headers: {
@@ -113,31 +272,24 @@ export function createClient(input: {
       })
       const json = (await tokens.json()) as any
       if (!tokens.ok) {
-        console.error(json)
-        throw new InvalidAuthorizationCodeError()
+        return {
+          err: new InvalidAuthorizationCodeError(),
+        }
       }
       return {
-        access: json.access_token as string,
-        refresh: json.refresh_token as string,
+        err: false,
+        tokens: {
+          access: json.access_token as string,
+          refresh: json.refresh_token as string,
+        },
       }
     },
     async refresh(
       refresh: string,
-      opts?: {
-        access?: string
-      },
-    ): Promise<
-      | {
-          err?: never
-          tokens?: Tokens
-        }
-      | {
-          err: InvalidRefreshTokenError | InvalidAccessTokenError
-        }
-    > {
+      opts?: RefreshOptions,
+    ): Promise<RefreshSuccess | RefreshError> {
       if (opts && opts.access) {
         const decoded = decodeJwt(opts.access)
-        console.log(decoded)
         if (!decoded) {
           return {
             err: new InvalidAccessTokenError(),
@@ -145,7 +297,9 @@ export function createClient(input: {
         }
         // allow 30s window for expiration
         if ((decoded.exp || 0) > Date.now() / 1000 + 30) {
-          return {}
+          return {
+            err: false,
+          }
         }
       }
       const tokens = await f(issuer + "/token", {
@@ -165,6 +319,7 @@ export function createClient(input: {
         }
       }
       return {
+        err: false,
         tokens: {
           access: json.access_token as string,
           refresh: json.refresh_token as string,
@@ -174,27 +329,8 @@ export function createClient(input: {
     async verify<T extends SubjectSchema>(
       subjects: T,
       token: string,
-      options?: {
-        refresh?: string
-        issuer?: string
-        audience?: string
-        fetch?: typeof fetch
-      },
-    ): Promise<
-      | {
-          err?: undefined
-          tokens?: Tokens
-          subject: {
-            [type in keyof T]: {
-              type: type
-              properties: v1.InferOutput<T[type]>
-            }
-          }[keyof T]
-        }
-      | {
-          err: InvalidRefreshTokenError | InvalidAccessTokenError
-        }
-    > {
+      options?: VerifyOptions,
+    ): Promise<VerifyResult<T> | VerifyError> {
       const jwks = await getJWKS()
       try {
         const result = await jwtVerify<{
@@ -209,6 +345,7 @@ export function createClient(input: {
         ].validate(result.payload.properties)
         if (!validated.issues && result.payload.mode === "access")
           return {
+            aud: result.payload.aud as string,
             subject: {
               type: result.payload.type,
               properties: validated.value,
@@ -234,7 +371,6 @@ export function createClient(input: {
           verified.tokens = refreshed.tokens
           return verified
         }
-        console.error(e)
         return {
           err: new InvalidAccessTokenError(),
         }