+++ packages/openauth/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @openauthjs/openauth
 
+## 0.2.5
+
+### Patch Changes
+
+- 8d6a243: fix: eliminate OTP bias and timing attack vulnerability
+- 873d1af: support specifying granular ttl for access/refresh token
+
+## 0.2.4
+
+### Patch Changes
+
+- 8b5f490: feat: Add copy customization to Code UI component
+
+## 0.2.3
+
+### Patch Changes
+
+- 80238de: return aud field when verifying token
+
+## 0.2.2
+
+### Patch Changes
+
+- 6da8647: fix copy for code resend
+
+## 0.2.1
+
+### Patch Changes
+
+- 83125f1: Remove predefined scopes from Spotify adapter to allow user-defined scopes
+
+## 0.2.0
+
+### Minor Changes
+
+- 8c3f050: BREAKING CHANGE: `client.exchange` and `client.authorize` signatures have changed.
+
+  `client.exchange` will now return an `{ err, tokens }` object. check `if (result.err)` for errors.
+  `client.authorize` now accepts `pkce: true` as an option. it is now async and returns a promise with `{ challenge, url}`. the `challenge` contains the `state` and `verifier` if using `pkce`
+
+  all exchanges have been updated to reflect this if you would like to reference
+
+### Patch Changes
+
+- 0f93def: refactor: update storage adapters to use Date for expiry
+
+## 0.1.2
+
+### Patch Changes
+
+- 584728f: Add common ColorScheme
+- 41acdc2: ui: missing copy in password.tsx
+- 2aa531b: Add GitHub Actions workflow for running tests
+
+## 0.1.1
+
+### Patch Changes
+
+- 04cd031: if only single provider is configured, skip provider selection
+
 ## 0.1.0
 
 ### Minor Changes